NANCY: Fix signed/unsigned warnings
Introduced with commit 6bf575df250b31c454667a4bcd343ba7da5be5ca

+++ engines/nancy/state/scene.h
@@ -137,7 +137,7 @@ public:
 	byte hasItem(int16 id) const;
 	byte getItemDisabledState(int16 id) const { return _flags.disabledItems[id]; }
 	void setItemDisabledState(int16 id, byte state) {
-		if (id < _flags.disabledItems.size())
+		if ((uint16)id < _flags.disabledItems.size())
 			_flags.disabledItems[id] = state;
 	}
